@@ -9,7 +9,7 @@
     }
   ],
   "require": {
-    "php": ">=8.0",
+    "php": ">=8.2",
     "fond-of-oryx/representative-company-user": "^1.3.0",
     "fond-of-oryx/representative-company-user-extension": "^1.1.0",
     "fond-of-oryx/representative-company-user-rest-api-permission": "^1.1.0",
@@ -31,7 +31,7 @@
     "psr-4": {
       "FondOfOryx\\": "tests/FondOfOryx/",
       "Generated\\": "src/Generated/",
-      "Orm\\Zed\\": "src/Orm/Zed/"
+      "Orm\\": "src/Orm/"
     }
   },
   "config": {
